+++ CHANGES.md
@@ -14,6 +14,8 @@ This will be the initial release of `coqffi`.
   with the `impure` attribute)
 - **Feature:** Add a new feature to assume all OCaml values of
   the input OCaml module interface are pure (`-fpure-module`)
+- **Feature:** Add `Seq.t` to the list of the primitive types
+  supported by `coqffi`
 - **Fix:** Make sure module names are always capitalized
 
 ## `coqffi.1.0.0~beta2`